----------------------------------------------------------------------

FAIL: test_lxu_cache_lookup (__main__.SplitTableBatchedEmbeddingsTest)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "split_table_batched_embeddings_test.py", line 3994, in test_lxu_cache_lookup
    dtype=torch.int,
AssertionError: False is not true

----------------------------------------------------------------------
Ran 35 tests in 759.368s

FAILED (failures=1)

+++ bench/EmbeddingSpMDMBenchmark.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) Facebook, Inc. and its affiliates.
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
  * All rights reserved.
  * This source code is licensed under the BSD-style license found in the
  * LICENSE file in the root directory of this source tree.
@@ -58,7 +58,7 @@ void run_benchmark(
 
   vector<float> embedding_table(num_rows * embedding_dim);
   normal_distribution<float> embedding_distribution;
-  for (int i = 0; i < embedding_table.size(); ++i) {
+  for (size_t i = 0; i < embedding_table.size(); ++i) {
     embedding_table[i] = embedding_distribution(generator);
   }
   vector<float16> embedding_table_fp16;
@@ -273,7 +273,7 @@ void run_benchmark(
           assert(
               false && "ERROR: refernce impl and JIT imp did not both succeed");
         } else if (success) {
-          for (int i = 0; i < output.size(); ++i) {
+          for (size_t i = 0; i < output.size(); ++i) {
             assert(output[i] == output_ref[i]);
             if (output[i] != output_ref[i]) {
               cout << i << " " << output[i] << " " << output_ref[i] << endl;
